+++ app/component/manuscript/manuscript-item/manuscript-item.js
@@ -2,7 +2,7 @@
 
 module.exports = {
   template: require('./manuscript-item.html'),
-  controller: ['$log', 'manuscriptService', ManuscriptItemController],
+  controller: ['$log', '$window', 'manuscriptService', ManuscriptItemController],
   controllerAs: 'manuscriptItemCtrl',
   bindings: {
     manuscript: '<',
@@ -10,7 +10,7 @@ module.exports = {
   }
 };
 
-function ManuscriptItemController($log, manuscriptService) {
+function ManuscriptItemController($log, $window, manuscriptService) {
   $log.debug('ManuscriptItemController');
 
   this.changeEdit = {
@@ -19,6 +19,18 @@ function ManuscriptItemController($log, manuscriptService) {
 
   this.deleteManuscript = function() {
     $log.debug('manuscriptItemCtrl.deleteManuscript()');
+
+    let confirmed = $window.confirm('Are you sure that you want to delete this chapter?');
+
+    if (confirmed) {
+      $window.alert('This chapter has been deleted.');
+    }
+
+    if (!confirmed) {
+      $log.log('chapter not deleted');
+      return false;
+    }
+
     manuscriptService.deleteManuscript(this.manuscript)
     .then(() => {
       $log.log('manuscript deleted');
